Move color components to packages directory and update references
