+++ src/blocks/advanced-btn/edit.js
@@ -11,8 +11,8 @@ import IconRender from '../../components/icons/icon-render';
 import TypographyControls from '../../components/typography/typography-control';
 import BoxShadowControl from '../../components/common/box-shadow-control';
 import WebfontLoader from '../../components/typography/fontloader';
-import KadenceColorOutput from '../../components/color/kadence-color-output';
-import PopColorControl from '../../components/color/pop-color-control';
+import { KadenceColorOutput } from '@kadence/helpers';
+import { PopColorControl } from '@kadence/components';
 import classnames from 'classnames';
 import ButtonStyleCopyPaste from './copy-paste-style';
 import flow from 'lodash/flow';
