+++ src/blocks/row-layout/edit.js
@@ -24,7 +24,7 @@ import ContainerDimensions from 'react-container-dimensions';
  */
 import KadenceRange from '../../components/range/range-control';
 import ResponsiveControl from '../../components/responsive/responsive-control';
-import KadenceColorOutput from '../../components/color/kadence-color-output';
+import { KadenceColorOutput } from '@kadence/helpers';
 import MeasurementControls from '../../components/measurement/measurement-control';
 import { PopColorControl } from '@kadence/components';
 import KadenceRadioButtons from '../../kadence-radio-buttons';
