+++ src/blocks/row-layout/save.js
@@ -4,7 +4,7 @@
 
 import classnames from 'classnames';
 import times from 'lodash/times';
-import KadenceColorOutput from '../../components/color/kadence-color-output';
+import { KadenceColorOutput } from '@kadence/helpers';
 import { __ } from '@wordpress/i18n';
 const {
 	Component,
