+++ src/blocks/tabs/edit.js
@@ -20,7 +20,7 @@ import MeasurementControls from '../../measurement-control';
 import filter from 'lodash/filter';
 import IconControl from '../../components/icons/icon-control';
 import IconRender from '../../components/icons/icon-render';
-import KadenceColorOutput from '../../kadence-color-output';
+import { KadenceColorOutput } from '@kadence/helpers';
 import KadencePanelBody from '../../components/KadencePanelBody';
 import { PopColorControl } from '@kadence/components';
 /**
