+++ src/blocks/testimonials/block.js
@@ -8,7 +8,7 @@
 import icons from '../../icons';
 import times from 'lodash/times';
 import IconRender from '../../components/icons/icon-render';
-import KadenceColorOutput from '../../components/color/kadence-color-output';
+import { KadenceColorOutput } from '@kadence/helpers';
 /**
  * Import attributes
  */
