+++ src/blocks/testimonials/edit.js
@@ -29,7 +29,7 @@ import IconControl from '../../components/icons/icon-control';
 import IconRender from '../../components/icons/icon-render';
 import KadenceMediaPlaceholder from '../../kadence-media-placeholder';
 import { PopColorControl } from '@kadence/components';
-import KadenceColorOutput from '../../components/color/kadence-color-output';
+import { KadenceColorOutput } from '@kadence/helpers';
 import ResponsiveRangeControls from '../../components/range/responsive-range-control';
 import ResponsiveMeasuremenuControls from '../../components/measurement/responsive-measurement-control';
 import KadencePanelBody from '../../components/KadencePanelBody';
