--- src/components/color/pop-color-editor.css
@@ -1,156 +1,0 @@
-/**
- * #.# Editor Styles
- *
- * CSS for just Backend enqueued after style.scss
- * which makes it higher in priority.
- */
-span.color-indicator-icon {
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0; }
-  span.color-indicator-icon .dashicons-admin-site {
-    position: absolute;
-    left: 6px;
-    top: 6px;
-    color: #444;
-    background: white;
-    width: 16px;
-    font-size: 14px;
-    height: 16px;
-    border: 1px solid white;
-    border-radius: 100%;
-    box-sizing: border-box;
-    padding-left: 1px;
-    padding-top: 1px; }
-  span.color-indicator-icon svg:not(.dashicon) {
-    width: 16px;
-    height: 16px;
-    top: 5px;
-    position: absolute;
-    left: 6px; }
-
-.components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item.is-active + .dashicons-admin-site {
-  position: absolute;
-  left: 6px;
-  top: 6px;
-  width: 18px;
-  font-size: 18px;
-  height: 18px; }
-
-.new-kadence-advanced-colors-pop.components-popover.kt-popover-color .components-color-palette {
-  padding: 2px 0 10px;
-  display: flex;
-  justify-content: flex-start;
-  flex-wrap: wrap;
-  min-width: 290px; }
-
-.new-kadence-advanced-colors-pop.components-popover.kt-popover-color .components-color-palette__item-wrapper {
-  margin-right: 2px;
-  margin-left: 2px;
-  margin-bottom: 4px;
-  transform: scale(1); }
-
-.new-kadence-advanced-colors-pop.components-popover.kt-popover-color .components-color-palette__item-wrapper:hover {
-  transform: scale(1.1); }
-
-.components-popover.kt-popover-color.new-kadence-advanced-colors-pop .components-popover__content {
-  padding: 10px 10px 0; }
-
-.components-popover.kt-popover-color.new-kadence-advanced-colors-pop .components-button.components-color-palette__item.is-active:not(:disabled):not([aria-disabled=true]):not(.is-secondary):not(.is-primary):not(.is-tertiary):not(.is-link):hover {
-  box-shadow: inset 0 0 0 4px; }
-
-.inline-color-popup-inner-wrap.block-editor-block-toolbar {
-  display: block;
-  border: 0;
-  box-shadow: none; }
-
-.components-popover.kt-popover-color > .components-popover__content > .components-base-control {
-  padding: 0 10px; }
-
-.kt-block-default-palette .components-color-palette, .components-popover.kt-popover-color .components-color-palette {
-  margin-right: 0;
-  width: auto; }
-
-.kt-block-defaults-modal .components-panel__body .components-base-control.components-color-picker__inputs-field {
-  margin-bottom: 0; }
-
-.components-popover.kt-popover-color .components-color-palette__item-wrapper, .kt-color-popover-container .components-color-palette__item-wrapper {
-  display: inline-block;
-  height: 28px;
-  width: 28px;
-  margin-right: 7px;
-  margin-left: 7px;
-  margin-bottom: 14px;
-  vertical-align: top;
-  transform: scale(1);
-  transition: transform .1s ease;
-  background-image: linear-gradient(45deg, #ddd 25%, transparent 0), linear-gradient(-45deg, #ddd 25%, transparent 0), linear-gradient(45deg, transparent 75%, #ddd 0), linear-gradient(-45deg, transparent 75%, #ddd 0);
-  background-size: 10px 10px;
-  background-position: 0 0, 0 5px, 5px -5px, -5px 0;
-  border-radius: 50%; }
-  .components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item {
-    display: inline-block;
-    vertical-align: top;
-    height: 100%;
-    width: 100%;
-    border: none;
-    border-radius: 50%;
-    background: transparent;
-    box-shadow: inset 0 0 0 14px;
-    transition: box-shadow .1s ease;
-    cursor: pointer; }
-  .components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item:after, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item:after {
-    content: "";
-    position: absolute;
-    top: -1px;
-    left: -1px;
-    bottom: -1px;
-    right: -1px;
-    border-radius: 50%;
-    box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.2);
-    border: 1px solid transparent; }
-  .components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item.is-active, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item.is-active {
-    box-shadow: inset 0 0 0 4px;
-    position: relative;
-    z-index: 1; }
-  .components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item.is-active + .dashicons-saved, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item.is-active + .dashicons-saved {
-    position: absolute;
-    left: 4px;
-    top: 4px; }
-  .components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item:focus, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item:focus {
-    outline: none; }
-  .components-popover.kt-popover-color .components-color-palette__item-wrapper .components-color-palette__item:focus:after, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item:focus:after {
-    content: "";
-    border: 2px solid #606a73;
-    width: 32px;
-    height: 32px;
-    position: absolute;
-    top: -2px;
-    left: -2px;
-    border-radius: 50%;
-    box-shadow: inset 0 0 0 2px #fff; }
-
-.kt-block-default-palette .components-color-palette__item-wrapper, .kt-color-popover-container .components-color-palette__item-wrapper {
-  display: inline-block;
-  height: 28px;
-  width: 28px;
-  margin-right: 14px;
-  margin-bottom: 14px;
-  vertical-align: top;
-  transform: scale(1.1);
-  transition: transform .1s ease; }
-  .kt-block-default-palette .components-color-palette__item-wrapper .components-color-palette__item:after, .kt-color-popover-container .components-color-palette__item-wrapper .components-color-palette__item:after {
-    content: "";
-    position: absolute;
-    top: -1px;
-    left: -1px;
-    bottom: -1px;
-    right: -1px;
-    border-radius: 50%;
-    box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.2);
-    border: 1px solid transparent; }
-
-.kt-block-default-palette .components-color-palette__item-wrapper:hover, .kt-color-popover-container .components-color-palette__item-wrapper:hover {
-  transform: scale(1.1); }
