--- src/components/color/pop-color-editor.scss
@@ -1,155 +1,0 @@
-/**
- * #.# Editor Styles
- *
- * CSS for just Backend enqueued after style.scss
- * which makes it higher in priority.
- */
- .kadence-pop-color-control {
-	.kadence-pop-color-container {
-		display: flex;
-		align-items: center;
-	}
-	.kadence-beside-label {
-		flex-grow: 1;
-		font-weight: 400;
-	}
-	.kadence-pop-color-clear {
-		background: transparent;
-		padding: 4px;
-		height: auto;
-		box-shadow: none;
-		border: 1px solid transparent;
-		margin-right: 4px;
-		line-height: 1;
-		span.dashicon {
-			font-size: 16px;
-			width: 16px;
-			height: 16px;
-		}
-	}
-	.kadence-pop-color-popovers {
-		display: flex;
-		align-items: center;
-	}
-	.single-pop-color {
-		margin-left: 8px;
-	}
-	.single-pop-color:first-child {
-		margin-left: 0;
-	}
-	.kadence-pop-color-icon-indicate {
-		height: auto;
-		position: relative;
-		transform: scale(1);
-		transition: transform .1s ease;
-		border-radius: 50%;
-		padding: 0;
-		&.kadence-has-alpha {
-			background-image: linear-gradient(
-		45deg, #ddd 25%, transparent 0),linear-gradient(
-		-45deg, #ddd 25%, transparent 0),linear-gradient(
-		45deg, transparent 75%, #ddd 0),linear-gradient(
-		-45deg, transparent 75%, #ddd 0);
-			background-size: 10px 10px;
-			background-position: 0 0,0 5px,5px -5px,-5px 0;
-		}
-		.kadence-pop-color-indicate {
-			width: 28px;
-			height: 28px;
-			border-radius: 50%;
-			margin: 0;
-		}
-	}
-}
-.components-popover.kadence-pop-color-popover {
-	.components-popover__content {
-		padding: 10px 10px 0;
-	}
-	.kadence-pop-color-palette-swatches {
-		padding: 2px 0 10px;
-		display: flex;
-		justify-content: flex-start;
-		flex-wrap: wrap;
-		min-width: 290px;
-		.kadence-color-palette__item-wrapper {
-			margin-right: 2px;
-			margin-left: 2px;
-			margin-bottom: 4px;
-			transform: scale(1);
-			display: inline-block;
-			height: 28px;
-			width: 28px;
-			transition: transform .1s ease;
-			background-image: linear-gradient( 
-		45deg, #ddd 25%, transparent 0), linear-gradient( 
-		-45deg, #ddd 25%, transparent 0), linear-gradient( 
-		45deg, transparent 75%, #ddd 0), linear-gradient( 
-		-45deg, transparent 75%, #ddd 0);
-			background-size: 10px 10px;
-			background-position: 0 0, 0 5px, 5px -5px, -5px 0;
-			border-radius: 50%;
-			&:hover {
-				transform: scale(1.1);
-			}
-			.kadence-color-palette__item {
-				display: inline-block;
-				vertical-align: top;
-				height: 100%;
-				width: 100%;
-				border: none;
-				border-radius: 50%;
-				background: transparent;
-				box-shadow: inset 0 0 0 14px;
-				transition: box-shadow .1s ease;
-				cursor: pointer;
-				&.is-active {
-					box-shadow: inset 0 0 0 4px;
-					position: relative;
-					z-index: 1;
-				}
-				&.is-active + .dashicons-admin-site {
-					position: absolute;
-					left: 6px;
-					top: 6px;
-					width: 18px;
-					font-size: 18px;
-					height: 18px;
-					z-index: 1;
-				}
-				&:after {
-					content: "";
-					position: absolute;
-					top: -1px;
-					left: -1px;
-					bottom: -1px;
-					right: -1px;
-					border-radius: 50%;
-					box-shadow: inset 0 0 0 1px rgba(0,0,0,.2);
-					border: 1px solid transparent;
-				}
-			}
-		}
-	}
-}
-span.color-indicator-icon {
-	position: absolute;
-    left: 0;
-    right: 0;
-    top: 0;
-    bottom: 0;
-	.dashicons-admin-site {
-		position: absolute;
-		left: 6px;
-		top: 6px;
-		color: #444;
-		background: #fff;
-		width: 16px;
-		font-size: 14px;
-		height: 16px;
-		border: 1px solid #fff;
-		border-radius: 100%;
-		box-sizing: border-box;
-		padding-left: 1px;
-		padding-top: 1px;
-	}
-}
