+++ src/packages/components/src/index.js
@@ -3,3 +3,4 @@ export { default as SinglePopColorControl } from './single-pop-color-control';
 export { default as ColorIcons } from './color-icons';
 export { default as ColorPicker } from './color-picker';
 export { default as ColorFields } from './color-fields';
+export { default as InlinePopColorControl } from './inline-pop-color';
