+++ src/packages/helpers/src/index.js
@@ -1,2 +1,4 @@
 export { default as getPreviewSize } from './get-preview-size';
 export { default as hexToRGBA } from './hex-to-rgba';
+export { default as KadenceColorOutput } from './kadence-color-output';
+export { default as DeprecatedKadenceColorOutput } from './deprecated-kadence-color-output';
