+++ src/packages/helpers/src/kadence-color-output/index.js
@@ -4,12 +4,12 @@
  * @returns {string} with var if needed.
  */
 /* global kadence_blocks_params */
-import hexToRGBA from './hex-to-rgba';
+import { hexToRGBA } from '@kadence/helpers';
 // eslint-disable-next-line camelcase
 export default function KadenceColorOutput( string, opacity = null ) {
 	if ( string && string.startsWith( 'palette' ) ) {
 		string = 'var(--global-' + string + ')';
-	} else if ( opacity !== null && ! isNaN( opacity ) && undefined !== string && '' !== string ) {
+	} else if ( opacity !== null && ! isNaN( opacity ) && 1 !== Number( opacity ) && undefined !== string && '' !== string ) {
 		string = hexToRGBA( string, opacity );
 	}
 	return string;
